[R] Change `*let` and `*and` into function, add `*or`
Form returned by optimizer was bit unclear on what it does for those who
aren't familiar with macro `*let` and `*and`.

    (*let (#1=#:object X)
      (*and ...))

Now they are just a function instead of macro and return its "expansion"
when called in optimization. So the form should be combination of stan-
dard operators and easy to read for anyone who are familiar with Common
Lisp. (esp. in cases that `let` or `and` removed completely, it turns
out to be very simple form.)